+++ src/main/java/randoop/condition/specification/OperationSpecification.java
@@ -95,9 +95,9 @@ public class OperationSpecification {
         operation,
         identifiers,
         // These cannot be Collections.emptyList() because the fields are mutable
-        new ArrayList<Precondition>(),
-        new ArrayList<Postcondition>(),
-        new ArrayList<ThrowsCondition>());
+        new ArrayList<>(),
+        new ArrayList<>(),
+        new ArrayList<>());
   }
 
   /**