+++ src/main/java/randoop/operation/ArrayCreation.java
@@ -68,6 +68,7 @@ public class ArrayCreation extends CallableOperation {
   // The argument array contains a single Integer.
   public ExecutionOutcome execute(Object[] input) {
     assert input.length == 1 : "requires array dimension as input";
+    @SuppressWarnings("PMD.UnnecessaryBoxing") // bug in PMD
     int length = ((Integer) input[0]).intValue();
     long startTimeMillis = System.currentTimeMillis();
     Object theArray = Array.newInstance(this.componentType.getRuntimeClass(), length);